Adjust buffer size properly
If you're creating a char buffer to which you're planning to strcpy stuff, you
need to take into account in its length space for the terminating null
character.

In this case, we sometimes strcpy "YES" there, so we need it to be of size 4.

While I'm fixing this here, I'd bet:
1) we have this same issue in other places (size 3 buffers for noyes results);
2) we have a similar issue with other buffers;